+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/CushionSpell.java
@@ -13,7 +13,6 @@ public class CushionSpell extends Spell
 {
 	private int cushionWidth = 3;
 	private int cushionHeight = 4;
-	private int airBubble = 3;
 	
 	@Override
 	public boolean onCast(String[] parameters) 
@@ -30,20 +29,17 @@ public class CushionSpell extends Spell
 		castMessage(player, "Happy landings");
 		
 		BlockList cushionBlocks = new BlockList();
-		cushionBlocks.setTimeToLive(10000);
-		
-		BlockList airBlocks = new BlockList();
-		airBlocks.setTimeToLive(500);
-		airBlocks.setRepetitions(30);
+		cushionBlocks.setTimeToLive(7000);
+		spells.disablePhysics(8000);
 		
 		int bubbleStart = -cushionWidth  / 2;
 		int bubbleEnd = cushionWidth  / 2;
 		
-		for (int dx = bubbleStart - airBubble; dx < bubbleEnd + airBubble; dx++)
+		for (int dx = bubbleStart; dx < bubbleEnd; dx++)
 		{
-			for (int dz = bubbleStart - airBubble ; dz < bubbleEnd + airBubble; dz++)
+			for (int dz = bubbleStart ; dz < bubbleEnd; dz++)
 			{
-				for (int dy = -airBubble; dy < cushionHeight + airBubble; dy++)
+				for (int dy = 0; dy < cushionHeight; dy++)
 				{
 					int x = targetFace.getX() + dx;
 					int y = targetFace.getY() + dy;
@@ -51,29 +47,15 @@ public class CushionSpell extends Spell
 					Block block = craftWorld.getBlockAt(x, y, z);
 					if (block.getType() == Material.AIR)
 					{
-						if (dx <= bubbleStart || dx >= bubbleEnd || dz <= bubbleStart || dz >= bubbleEnd || dy <= 0)
-						{
-							airBlocks.add(block);
-						}
-						else
-						{
-							cushionBlocks.add(block);
-							block.setType(Material.STATIONARY_WATER);
-						}
+    					cushionBlocks.add(block);
+    					block.setType(Material.STATIONARY_WATER);
 					}
 				}
 			}
 		}
 	
 		spells.scheduleCleanup(cushionBlocks);
-		spells.scheduleCleanup(airBlocks);
 	
-		// Schedule an additional later cleanup, to cleanup water spillage
-		BlockList delayedCleanup = new BlockList(cushionBlocks);
-		delayedCleanup.setTimeToLive(15000);
-		
-		spells.scheduleCleanup(delayedCleanup);
-		
 		return true;
 	}
 
@@ -100,7 +82,6 @@ public class CushionSpell extends Spell
 	{
 		cushionWidth = properties.getInteger("spells-cushion-width", cushionWidth);
 		cushionHeight = properties.getInteger("spells-cushion-height", cushionHeight);
-		airBubble = properties.getInteger("spells-air-bubble-thickness", airBubble);
 	}
 
 	@Override